Respond to Cache-Control headers (PP-2380) (#2417)
## Description

- Remove the internal CM authentication document cache
- We are caching the authentication documents in cloudfront, so no need
to have an additional cache within the CM
- Add a new decorator that responds to cache control headers and apply
it to `authentication_document` and `groups` urls.

## Motivation and Context

This PR is the first step to resolve both PP-2380 and PP-2398. When we
know we need to load a fresh document without caching the client will be
responsible for setting the headers appropriately.

This will need 3 additional PRs to fully function:
- Library registry to make sure it sets the headers when requesting auth
document
  - PR: https://github.com/ThePalaceProject/library-registry/pull/821 
- Admin UI to make sure it sets the headers when requesting groups feed
  - https://github.com/ThePalaceProject/web-opds-client/pull/44
  - https://github.com/ThePalaceProject/circulation-admin/pull/157
- Cloudfront to make sure our cache keys take into account the
cache-control header
  - PR: https://github.com/ThePalaceProject/hosting-playbook/pull/811

However this PR can be reviewed and merged before those follow ups are
in place.

## How Has This Been Tested?

- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.