[PP-1482] migrate playtime summary scripts to celery (#2415)
## Description

<!--- Describe your changes -->
This PR moves the meat of the playtime summation and report logic from
the scripts into celery tasks. I left the scripts endpoints in place to
facilitate manually generating reports or kicking off the summation
task.

Virtually all the code changes here are simply moves: most of the logic
and test code hasn't really changed.

I left a few comments and a couple of questions for consideration.
## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-1482
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have updated the documentation accordingly.
- [ ] All new and existing tests passed.