Configurable SIP2 timeout (PP-2418) (#2424)
## Description

Allow configuration of SIP2 connection timeout.

## Motivation and Context

We set the SIP2 connection timeout fairly low, because it has impacts on
all the performance for all libraries on a CM. A single slow library,
can slow the rest of them down. Libraries with slow authentication
should cause less of a performance impact now that we have token auth
rolled out everywhere, as we don't have to call out to the ILS as often,
so I think adding a configuration that allows us to increase the timeout
makes sense.

## How Has This Been Tested?

- Added tests
- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.