Mark LCP content as deliverable (PP-2334) (#2429)
## Description

Add LCP to the list of deliverable formats in the mobile clients.

## Motivation and Context

Eventually, I think this property should be removed, as we have
different clients, and they can deliver different formats. For example
the web catalog can only deliver some formats, and the mobile apps might
differ for some time in their ability to deliver a format.

However we need bigger changes for that, and this property is still used
in one of the importers and the custom list code. So I'm going to leave
it for now, but update it so that LCP is marked as supported.

## How Has This Been Tested?

- Running tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.