[PP-1471] convert novelist script to celery tasks (#2443)
## Description
This update replaces the implementation of the NovelistSnapshotScript to
spawn a celery task for each library we want to update. On the celery
side, there are two tasks. One spawns tasks for each library while the
other handles the novelist updates by library.

Additionally I removed the cron job from
docker/services/cron/circulation and moved it to a celery scheduled task
executing at the same time: ie Midnight on Sunday.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-1471

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests added.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.