Optimize LibraryAnnotator NoveList config check. (PP-2491) (#2449)
- Lazy load and cache the value of `NoveList.is_configured` in the
`LibraryAnnotator`. The previous behavior was to compute this value
multiple times, once for every feed entry.
- Added test to ensure that `NoveList.is_configured` is called only once
per feed generation, regardless of how many items are in the feed.
Without the change this test was failing. After the change, it passes.

Computing the value of `NoveList.is_configured` was contributing to
slowness (1/5 to 1/4 of response time in my testing) when building our
OPDS feeds because a database query is being run every time we check to
see if NoveList is configured for a given library. However, the value
should not change within a single feed build, so computing once and
caching the result makes sense here.

- Added new test.
- All tests pass locally.
- Tested with profiling in local development environment.
- After this change, calls to `NoveList.is_configured` were no longer
contributing significantly to the response time and the response time
improved.
- [CI
tests](https://github.com/ThePalaceProject/circulation/actions/runs/14845548918)
pass.

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.