This update fixes a bug that was causing google drive failures when t… (#2450)
…he GOOGLE_DRIVE_PARENT_FOLDER_ID refers to a shared drive.

## Description
The production reports were not being uploaded to the shared google
drive.
<!--- Describe your changes -->

## Motivation and Context

Tara Carpenter was not seeing the files land in the EBCE Partner Success
Team shared drive.

<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I generated a report from my local instance using the production google
drive credentials and parent folder id and verified that the report
landed. I also tested that it still works when the folder id is not in a
shared drive.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.