[PP-2504] change playtime report name (#2456)
## Description

Now that we are writing production reports to google drive, Tara and I
discovered a couple of minor bugs that were resulting in reports not
landing in quite the right place with the right file naming conventions.
This PR addresses those mismatches.

## Motivation and Context
https://ebce-lyrasis.atlassian.net/browse/PP-2504
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests updated and manually tested in the production google drive.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.