Add pydantic serialization / deserialization to celery (kombu) (PP-2468) (#2461)
## Description

This PR allows you to pass pydantic models into celery function calls,
and have them automatically serialized and deserialized, so they are
transparently passed into the function.

This PR also updates the marc exporter task to use this new
functionality. Previously it was manually dumping the loading the
pydantic models it was using.

## Motivation and Context

As part of PP-2468 we are going to be passing more pydantic models into
celery tasks, so this is an effort to make that easier.

## How Has This Been Tested?

- Added unit test
- Converted marc exporter to use new functionality

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.