Use pydantic for Metadata and CirculationData (PP-2466) (#2458)
## Description

Update `Metadata` and `CirculationData` to use pydantic for parsing and
serialization.

Note: Unlike the conversions done in
https://github.com/ThePalaceProject/circulation/pull/2448 the top level
"containers" are not immutable and hashable, so they can be built up and
modified in the importers.

## Motivation and Context

Follow up to: https://github.com/ThePalaceProject/circulation/pull/2448

Allow us to serialize `Metadata` and `CirculationData` objects, so that
we can pass them into Celery tasks.

The biggest change that was necessary to convert these classes was to
standardize how we are passing the `data_source` and
`primary_identifier`. Previously we were not very strict about what went
into these, and the constructor sorted it out. This updates the name of
these fields to make it clear what they are, and adds some methods to
load them from the database if necessary.

See: PP-2466

## How Has This Been Tested?

- Running unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.