Rename metadata_layer.Metadata to data_layer.BibliographicData (PP-2468) (#2463)
## Description

Rename the `metadata_layer` package to `data_layer` and rename
`Metadata` -> `BibliographicData`.

## Motivation and Context

This PR is based on the code review discussion in this PR
https://github.com/ThePalaceProject/circulation/pull/2458.

This one is a big one. `Metadata` is referenced all over the code base,
especially in some untyped parts, which makes renaming difficult. It is
also referenced in function and variable names, and in comments pretty
widely. I did my best to find all the references and clean them up as
part of this PR.

Some of this work is hard because we have `metadata` integrations, which
do still need metadata references in their names. So its not as easy as
doing a find and replace.

## How Has This Been Tested?

- Running unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.