Celery task to call BibliographicData.apply (PP-2468) (#2465)
## Description

- Add a new queue `apply`
- Add a new worker to consume from that queue
- Add a new task `bibliographic_apply` that calls
`BibliographicData.apply`

Note: Nothing is currently pushing data into this queue for the task,
this is just getting the groundwork setup. I will modify one of our
importers to push data into this queue in a follow up PR.

## Motivation and Context

JIRA: PP-2468

Getting tasks setup, so I can convert the OPDS importers to use Celery. 

This PR is based on
https://github.com/ThePalaceProject/circulation/pull/2463, so that one
should go in first.

## How Has This Been Tested?

- Added new tests for this case.

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.