Add celery tasks to mark unavailable identifiers (PP-2469) (#2468)
## Description

Adds a series of tasks meant to provide the foundation for provider
specific celery tasks that mark any identifiers not found in the
providers feed as unavailable. The intention is to compose these tasks
using celery workflows.

It also adds a new redis data structure to store sets of items, so we
are able to store sets of identifiers that persist over mutliple celery
task invocations.

Note: these tasks are not used anywhere yet. I'm planning to use them in
a follow up to convert the OPDS importer tasks.

## Motivation and Context

See PP-2469. This is building up some of the tools needed to convert the
OPDS importer tasks to Celery.

## How Has This Been Tested?

- New unit tests added to CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.