[PP-1792] add support for collection depriorization (#2475)
## Description

What I've done here is the following:
1. I added a configuration parameter for collections that allows the
user to specify the "lane_priority_level" which is on a scale of 1 to
10.
2. I updated the search index to include the
licensepools.lane_priority_level field.
3. I added `field_value_factor` to the featurability_scoring_functions
which simply adds the lane_priority_level to the score.
4. I added testing.

Note: the `field_value_factor` allows for both a `modifier` property as
well as a `factor`. The modifier is one of a number of math functions
you can apply to the value to manipulate the magnitude of the score. So
I opted for no modifier and no factor. To the extent that a user
deviates from the middle of the range (ie 5), so she will give or take
weight from the items in the collection. The far end of the ranges the
user can effectively trump the quality scores. When the differences
between colleciton priority levels are lower, the priority won't
necessarily trump the quality scores or the effects of the randomizing
function.

## Motivation and Context
Sometimes content from one collection can swamp a lane despite their
being other collections matching the lane criteria. This feature enables
users to specify a priority for the collection that applies to works
that are featured in lanes. For example a librarian may find that they
the lane is only displaying books from an open access collection but
s/he wants to give priority to the licensed content that is featured in
the lane. In this case, the user would increase or reduce collection
lane priorities according to the desired outcome.

https://ebce-lyrasis.atlassian.net/browse/PP-1792
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Manually and unit tested.

For manual testing you must do the following:
Starting conditions:
Have at least two collections. 
For a given lane, verify that works from both collections appear in the
lane.

1. change the collection lane priority levels according to your tests.
2. reindex the collections
3. validate that the ordering of the lane is what you would expect (ie
the collections with the highest priority appear first in the list and
those with the lowest appear at the end. If the spread between lowest
and highest is minimal - say you have a 6 and 4, you may not see all the
6's coming before the 4's. Quality may be a factor. The wider the
spread, the less quality will factor in.


<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.

---------

Co-authored-by: Tim DiLauro <tdilauro@users.noreply.github.com>