[PP-2555] Catch and log expected NYT celery exceptions. (#2488)
## Description
This PR ensures that exceptions in the NYT bestsellers update celery
task arising due to the absence of a NYT bestseller service are caught
and logged within the task.

## Motivation and Context
We don't want to see uncaught celery exceptions when a PM doesn't have
an NYT integration configured. A warning should be sufficient.

https://ebce-lyrasis.atlassian.net/browse/PP-2555
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
New unit test added.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.