Speed up celery tests by using session worker fixture (PP-2569) (#2491)
## Description

Use a Celery worker the persists for the whole session, rather then
starting and stopping a celery worker for each test.

## Motivation and Context

We used to be able to run the tests in 5 - 6 minutes, but we're up to
around 10 minutes now. Looking at whats slowing things down, its mostly
the Celery tests. The process of starting and stopping a Celery worker
for each of our Celery tests ends up slowing down out tests
significantly.

Celery provides a `celery_session_worker` fixture that only starts the
worker once, in a thread, and then uses that same worker for all the
tests. There is a bit more risk doing things this way, that state will
leak between tests. In practice though, I think this is reasonably
minimal, since Celery workers are supposed to be fairly stateless to
start with. They just process the jobs sent to them.

This change seems like it takes a couple minutes off the time is takes
to do a test run in CI.

## How Has This Been Tested?

- Running tests locally
- Running tests in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.