Create session fixture for services container (PP-2569) (#2492)
## Description

- Create a session fixture that contains the autospec mocks for our
services container, so we don't have to run `create_autospec` on every
test run.
- Refactor our services fixture, so it uses the new session fixture. 
- Remove a number of fixtures that were created for the sub-containers
that weren't very useful, and replace them with calls to the main
services fixture.

## Motivation and Context

`create_autospec` is surprisingly expensive to run, and with our current
services container, we have to run several times for almost every test
run. This ends up making out tests take longer.

I think the new structure where we don't have a bunch of different
services fixtures is a better design and it shaves a minute or so off
our tests runs.

## How Has This Been Tested?

- Tested locally
- Tested in CI

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.