[PP-2610] Ensure that published dates from OPDS2 imports contain no time info. (#2501)
when being imported into palace.

## Description
This PR addresses the import error we were seeing on some titles that
had time data in in the "published" field in the OPDS import. Pydantic
wasn't liking it:

```
pydantic_core._pydantic_core.ValidationError: 1 validation error for BibliographicData
published
  Datetimes provided to dates should have zero time - e.g. be exact dates [type=date_from_datetime_inexact, input_value=datetime.datetime(1970, 1... 43, tzinfo=TzInfo(UTC)), input_type=datetime]
    For further information visit https://errors.pydantic.dev/2.11/v/date_from_datetime_inexact
```


<!--- Describe your changes -->

## Motivation and Context
See above.  Or for more details - see the following link:
https://ebce-lyrasis.atlassian.net/browse/PP-2610
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I added a unit test the reproduce the failure. I fixed it using a
validator and confirmed that the fix prevents the error.

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.