[PP-1468] convert work classification to celery (#2500)
## Description
This PR migrates the `bin/work_presentation_editions` script and the
`bin/work_classification` script to the celery.

These scripts would run periodically to reclassify and recalculate work
presentation editions based on the state of the work coverage records.

This PR removes the WorkCoverageRecord and the underlying database
structures. Instead of manipulating a work's coverage records in order
to cause classification and presentation editions recalculations, we
simply queue the work id and the calculation policy in redis.
Periodically the `calculate_work_presentations` celery task wakes up and
processes any work ids that have been queued for recalculation.

## Motivation and Context
This ticket is part of a larger effort to move away from cron based
scripts towards an asynchronous event driven system for task processing.
The major improvement here is that we are no longer having to stuff an
ever-growing list of coverage records for every work in the system into
the database. Instead we dispense with the coverage records entirely and
instead use redis to track the list of items for update. This should
reduce the contention on the workcoveragerecords table which we believe
had a significant role in the ongoing problem of relatively frequent
dead locks.

https://ebce-lyrasis.atlassian.net/browse/PP-1468



<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests cover the functionality. Once merged I'll test it on
minotaur.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.