[PP-1479] convert saml scripts to celery (#2502)
## Description
This PR converts the saml_monitor script to a celery task. It runs on
the same schedule (ie once a day at 5am). It can be manually triggered
using the ./bin/update_saml_metadata script.

## Motivation and Context
This task is part of the larger effort to convert our scripts to use the
celery task processing framework.
https://ebce-lyrasis.atlassian.net/browse/PP-1479
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
New unit tests cover the new functionality.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.