Refactor MockHttpClient to be pytest fixture (#2525)
## Description

Refactor MockHttpClient to be pytest fixture. Previously it was a
standalone class, but it makes more sense for it to be a fixture that
other test functions can call in.

## Motivation and Context

This change was part of the work I'm doing in PP-2133. Breaking it out
to its own PR, since PP-2133 is getting rather large.

## How Has This Been Tested?

- Running unit tests

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.