Remove superseded license pool concept. (PP-2590) (#2526)
## Description

Removes the concept of a superseded ("superceded" [sic] in code) license
pool.

NB: The `LicensePool.superceded` column will be removed in a future
update, but all references to it have been removed.

## Motivation and Context

Before this change, Open Access license pools could be superseded by
another license pool. This caused problems if the _superseding_ license
pool was not associated with and active in all libraries that were
associated with the _superseded_ pool.

## How Has This Been Tested?

- Removed or reworked tests that used superseded.
- All tests pass locally.
- [CI
tests](https://github.com/ThePalaceProject/circulation/actions/runs/15682544711).

## Checklist

- N/A - I have updated the documentation accordingly.
- [X] All new and existing tests passed.