[PP-2523] Hide currently selected book titles (#2537)
## Description
This PR filters the related_books endpoint results such that the work
identified in the /related_books URI does not appear in the results (
since that would be redundant).

I've added a test that ensures the values passed via the
"work_ids_to_exclude" parameter of the `OPDSAcquisitionFeed.groups`
method are being filtered out of the generated OPDS acquisition feed.

I also updated the test of the `WorkController.related` method to ensure
that the work id is being passed to the `OPDSAcquisitionFeed.groups`
method.

## Motivation and Context
When a user is looking at the book details for Book A in the mobile
apps, we do not want Book A to appear in the related book results.
https://ebce-lyrasis.atlassian.net/browse/PP-2523
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
I tested it locally with manual testing, updated the unit tests, and
added a new test.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.