[PP-1469] convert work classify unchecked to celery (#2519)
## Description
This PR refactors the work classification for unchecked subjects
script's implementation into the celery framework, making use of the
redis set for storing work_id and calculation policy tuples for later
processing. The celery task scheduler kicks off the unchecked subject
processing once a day as before. I've also preserved the ability to
manually kick off the task via a script.


## Motivation and Context
This PR converts the work classification for unchecked subjects script
to use the celery task framework.

https://ebce-lyrasis.atlassian.net/browse/PP-1469
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have updated the documentation accordingly.
- [ ] All new and existing tests passed.