Allow axis to set format prioritities (PP-2133) (#2547)
## Description

Allow the Axis integration to set format priorities settings, so that we
are able to prioritize / de-prioritize the new AxisNow DRM
implementation.

Once this is rolled out and working, we will want to make sure AxisNow
is always prioritized, but for now adding these settings is helpful for
testing.

## Motivation and Context

Testing PP-2133

## How Has This Been Tested?

- Tested locally

## Checklist

- [X] I have updated the documentation accordingly.
- [X] All new and existing tests passed.