[PP-2608] Generate empty reports for data sources associated with ODL… (#2524)
… 2.0 and OPDS For Distributors based collections.

## Description
Before this PR, we were generating a list of data sources for which to
generate playtime reports solely based on whether or not there were
playtime summaries associated with those datasources in the desired time
range. As a result, reports were generated only if there was data to
report. However the PST wants to see an empty report in the case of no
data so that there is no question about whether a report was generated.

This change ensures that reports are generated for all Data soruces
associated with OPDS for Distributors and ODL 2.0 providers since these
protocols align with collections which we want to include in our
playtime reports.

## Motivation and Context

https://ebce-lyrasis.atlassian.net/browse/PP-2608
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Unit tests were updated to reflect changes.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] I have updated the documentation accordingly.
- [x] All new and existing tests passed.